requirements.txt
flask
gunicorn
flask_sqlalchemy
python-dotenv
psycopg2-binary
Flask-Babel
requests
supabase
